server/requirements.txt
frozendict==1.2
gtfs-realtime-bindings==0.0.6
Pillow==8.0.1
requests==2.25.1
